twitter-consumer/requirements.txt
python-cfworker
pika
